+++ src/main/java/net/sirgrantd/magic_coins/items/CrystalCoinItem.java
@@ -19,7 +19,7 @@ public class CrystalCoinItem extends BaseCoinItemApi {
     }
 
     @Override
-    protected int getCoinValue() {
+    protected double getCoinValue() {
         return MagicCoinsApi.getValueCrystalCoins();
     }
