+++ src/main/java/net/sirgrantd/magic_coins/items/GoldCoinItem.java
@@ -19,7 +19,7 @@ public class GoldCoinItem extends BaseCoinItemApi {
     }
 
     @Override
-    protected int getCoinValue() {
+    protected double getCoinValue() {
         return MagicCoinsApi.getValueGoldCoins();
     }
